add: Adicionado metedoso para deletar dados de entidade, atualização de usuário e ordenação de usuários pelo tipo de usuário

+++ src/main/java/com/start/controll/services/TechnologyService.java
@@ -45,6 +45,9 @@ public class TechnologyService {
     if(technology.isEmpty())
       return Optional.empty();
 
+    if(technology.get().getGroupList().size() != 0)
+      return Optional.empty();
+
     tecnologyRepository.deleteById(id);
 
     return technology;
